Ignore local config file
It is helpful to test locally with a local config file with the
following changes in the baseurl and url lines:
baseurl: ""
url:

Then, you can test with:
bundle exec jekyll serve --livereload --config _config_local.yml

We could either add that _config_local.yml to the repo or just keep it
locally and ask git to ignore it. I went to the last option.